Remove incorrect changes brought in from merges that shouldn't be in this PR
